refactor player
